+++ src/portal/components/Loading.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Row, Spin } from 'antd';
 
 const Loading = React.memo(() => (
-  <Row type="flex" justify="center">
+  <Row justify="center">
     <Spin />
   </Row>
 ));