+++ src/shared/components/layout/tags/RollBtn/shared.tsx
@@ -19,7 +19,7 @@ export function roll(requestStr: string) {
   try {
     let pattern = /(\d*)\s*d\s*(\d*)/gi;
 
-    requestStr = requestStr.replace(/[^\dd\+-\/\*]+/gi, ''); //去除无效或危险字符
+    requestStr = requestStr.replace(/[^\dd\+-\/\*\(\)]+/gi, ''); //去除无效或危险字符
     let express = requestStr.replace(pattern, function(tag, num, dice) {
       num = num || 1;
       dice = dice || 100;