+++ src/web/components/Alert.tsx
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 import { hideAlert } from '../../shared/redux/actions/ui';
 import './Alert.scss';
 import { TRPGState, TRPGDispatchProp } from '@redux/types/__all__';
+import { Button } from 'antd';
 
 interface Props extends TRPGDispatchProp {
   showAlert: boolean;
@@ -18,16 +19,19 @@ class Alert extends React.Component<Props> {
     const onConfirm = showAlertInfo.onConfirm;
     const onCancel = showAlertInfo.onCancel;
 
-    let cancelBtn;
+    let cancelBtn: ReactNode;
     if (onConfirm) {
       cancelBtn = (
-        <button
+        <Button
+          type="primary"
+          block={true}
+          size="large"
           onClick={() =>
             onCancel ? onCancel() : this.props.dispatch(hideAlert())
           }
         >
           取消
-        </button>
+        </Button>
       );
     }
 
@@ -35,18 +39,23 @@ class Alert extends React.Component<Props> {
       <div>
         <div className="header">{title || ''}</div>
         <div className="body">{content || '确认进行该操作?'}</div>
-        <button
-          onClick={() => {
-            if (!!onConfirm) {
-              onConfirm();
-            } else {
-              this.props.dispatch(hideAlert());
-            }
-          }}
-        >
-          {confirmTitle || '确认'}
-        </button>
-        {cancelBtn}
+        <div className="action">
+          {cancelBtn}
+          <Button
+            type="primary"
+            block={true}
+            size="large"
+            onClick={() => {
+              if (!!onConfirm) {
+                onConfirm();
+              } else {
+                this.props.dispatch(hideAlert());
+              }
+            }}
+          >
+            {confirmTitle || '确认'}
+          </Button>
+        </div>
       </div>
     );
   }